requirements/requirements.txt
-r base.txt
python-json-logger
gunicorn
langchain==0.1.10
fastapi==0.110.0
uvicorn==0.28.0
async-asgi-testclient==1.4.11
pytest-asyncio==0.23.5
sentry-sdk==1.41.0
redis==5.0.3
PyMuPDF==1.24.2
tiktoken==0.6.0
pdfplumber==0.11.0
langchain-openai==0.1.6
jsonschema==4.22.0
python-multipart==0.0.9
exa-py==1.0.12
streamlit==1.35.0
SQLAlchemy-Utils==0.41.2
openai==1.35.0
qdrant-client==1.9.2
motor==3.5.0